services/producer/requirements.txt
statsbombpy==1.17.0
confluent-kafka==2.6.1
orjson==3.10.15
python-dotenv==1.1.0
